refactor(plugin): remove delegator methods; switch callers to services; update tests and metrics; add breaking-changes docs
