+++ src/main/java/com/skyblockexp/ezeconomy/core/MessageProvider.java
@@ -134,8 +134,8 @@ public class MessageProvider {
         if (template == null) {
             template = fallbackConfig.getString("price_message_format", "{amount} {symbol}");
         }
-        String amountStr = plugin.formatAmountOnly(amount, currency);
-        String symbol = plugin.getCurrencySymbol(currency);
+        String amountStr = plugin.getCurrencyFormatter().formatAmountOnly(amount, currency);
+        String symbol = plugin.getCurrencyFormatter().getCurrencySymbol(currency);
         java.util.Map<String, String> placeholders = java.util.Map.of("amount", amountStr, "symbol", symbol == null ? "" : symbol);
         return format(template, placeholders);
     }
